Word select() bug fixed
I no longer assume that meaning is a must while translation is optional.
Both are now optional. The word.select() method has been updated to
reflect this change. The syntax is a bit ugly if you need translation
but not meaning (as that requires "undefined, number") but I would
assume that most of the time meaning is required as well.

I also considered splitting select into two methods, one for meaning and
translation respectively, but that leads to issues with determining
whether selection is still pending.