Reconfigure whether Database commands fail loudly
There's a lot of debate here, but I've settled it to be such that
methods which don't explicitly find the card will fail loudly. This is
an attempt to uphold the abstraction barrier of the CardDatabase
interface. We have to essentially use Card.Front as the key to the card,
 instead of some arbitrary number like in the case of Anki.